+++ engine/docs/changelog.md
@@ -1,4 +1,7 @@
 # TBD
+### Features
+* Try to pull `api-container` latest image before running the API container Docker container
+* Upgraded to Kurt Core 1.25.1, which add `com.kurtosistech.app-id` container label to all enclave containers
 
 # 0.1.0
 ### Features